refactor: 게임 진행 로직 BaseballGame 클래스에게 위임
